Reorganize models into packages. Update classes.puml. Use pydantic and inherit from BaseModel.
