+++ src/models/wikimedia/enum.py
@@ -1,0 +1,35 @@
+from enum import Enum
+
+
+class WikimediaLanguageCode(Enum):
+    BASQUE = "eu"
+    BENGALI = "bn"
+    BOKMÅL = "nb"
+    CZECH = "cs"
+    DANISH = "da"
+    ENGLISH = "en"
+    ESTONIAN = "et"
+    FRENCH = "fr"
+    GERMAN = "de"
+    HEBREW = "he"
+    LATIN = "la"
+    MALAYALAM = "ml"
+    RUSSIAN = "ru"
+    SWEDISH = "sv"
+
+
+class WikimediaLanguageQID(Enum):
+    BASQUE = "Q8752"
+    BENGALI = "Q9610"
+    BOKMÅL = "Q25167"
+    CZECH = "Q9056"
+    DANISH = "Q9035"
+    ENGLISH = "Q1860"
+    ESTONIAN = "Q9072"
+    FRENCH = "Q150"
+    GERMAN = "Q188"
+    HEBREW = "Q9288"
+    LATIN = "Q397"
+    MALAYALAM = "Q36236"
+    RUSSIAN = "Q7737"
+    SWEDISH = "Q9027"
